// src/interfaces/enums.ts
export enum RoleEnum {
  USER = 'user',
  ADMIN = 'admin',
  ROOT = 'root',
}

export enum FileTypesEnum {
  PDF = 'pdf',
  IMAGE = 'image',
}

export enum OrderStatusEnum {
  done = 'done',
  todo = 'todo',
  inProcess = 'inProcess',
}

export enum MaterialTypeEnum {
  metr = 'metr',
  metr2 = 'metr2',
  piece = 'piece',
}

export enum UserGenderEnum {
  male = 'male',
  female = 'female',
}

export enum UserPositionEnum {
  cashier = 'cashier',
  seamstress = 'seamstress',
  admin = 'admin',
}

export enum UserRankEnum {
  junior = 'junior',
  leading = 'leading',
  senior = 'senior',
}